Fixed loading new regions/cities
Made it so when you select a new country or region, the select element under it will be populated with new values (or no values if relevant) and reset to the first option. Before this, it would save the index of the select element, meaning even after swapping countries/regions, it would remain selecting the Nth option under it.